transformer-robot-controller/services/mir_controller/requirements.txt
loguru==0.5.3
PyYAML==5.3.1
httpx==0.0.0